Put connect screen code in a separate file
