+++ src/operator/gui/connect_screen_logic.py
@@ -1,0 +1,14 @@
+from PyQt6.QtWidgets import QMainWindow
+from gui.views.connect_screen import Ui_ConnectScreen
+
+# Screen for connecting to C2 Server
+class ConnectScreen(QMainWindow, Ui_ConnectScreen):
+    def __init__(self):
+        super().__init__()
+        self.setupUi(self)
+        self.connectScreenStyleSheet = self.loadStyleSheet()
+        self.setStyleSheet(self.connectScreenStyleSheet)
+
+    def loadStyleSheet(self):
+        remoteMachinesStyleSheet = open("gui/resources/stylesheets/connectScreen.css", "r")
+        return remoteMachinesStyleSheet.read()
